% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_correctly_formatted_dates.R
\name{are_correctly_formatted_dates}
\alias{are_correctly_formatted_dates}
\title{Are all the dates correctly formatted?}
\usage{
are_correctly_formatted_dates(date_strs)
}
\arguments{
\item{date_strs}{strings for dates}
}
\value{
a logical vector,
\code{TRUE} for each string that follows a \code{YYYY-MM-DD} format
}
\description{
Are all the dates correctly formatted?
}
\note{
this does not test the value of the strings
}
