% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_valid_links.R
\name{are_valid_links}
\alias{are_valid_links}
\title{Detect if the links are valid}
\usage{
are_valid_links(urls)
}
\arguments{
\item{urls}{one or more URLs}
}
\value{
TRUE if the link is valid, FALSE otherwise
}
\description{
Detect if the links are valid
}
\examples{
urls <- c(
  "https://nbisweden.github.io/SCoRe_user_doc",
  "https://nonsense.url"
)
are_valid_links(urls)
}
\seealso{
use \link{is_valid_link} for checking one link
}
