% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_valid_urls.R
\name{are_valid_urls}
\alias{are_valid_urls}
\title{Determine if the URLs are correctly formatted.}
\usage{
are_valid_urls(urls)
}
\arguments{
\item{urls}{one or more URLs}
}
\value{
logical vector with the same length as the URLs
}
\description{
It does not check if the URLs exist
}
