% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_courses_table.R
\name{check_courses_table}
\alias{check_courses_table}
\title{Check if the table is correctly formatted, will \link{stop} if not}
\usage{
check_courses_table(t)
}
\arguments{
\item{t}{a table}
}
\value{
Nothing. Will \link{stop} with a proper error message when
the table is not formed correctly
}
\description{
Check if the table is correctly formatted, will \link{stop} if not
}
