% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_providers_info_file.R
\name{check_providers_info_file}
\alias{check_providers_info_file}
\title{Check the file to provide a table with all of the providers' information}
\usage{
check_providers_info_file()
}
\value{
nothing, will \link{stop} if there is a problem with the file
}
\description{
Check the file to provide a table with all of the providers' information
}
