% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_enccs_dates.R
\name{convert_enccs_dates}
\alias{convert_enccs_dates}
\title{Convert an ECSS data to \code{YYYY-MM-DD} format}
\usage{
convert_enccs_dates(enccs_dates)
}
\arguments{
\item{enccs_dates}{a date in ENCCS formal}
}
\value{
a string in \code{YYYY-MM-DD} format
}
\description{
Convert an ECSS data to \code{YYYY-MM-DD} format
}
\examples{
convert_enccs_dates("Jul 10") # 2025-07-10
}
