% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_english_dates_to_iso_8601.R
\name{convert_english_dates_to_iso_8601}
\alias{convert_english_dates_to_iso_8601}
\title{Convert one or more dates written in English
to the ISO 8601 \code{YYYY-MM-DD} format}
\usage{
convert_english_dates_to_iso_8601(english_dates)
}
\arguments{
\item{english_date}{a date in Enlish}
}
\value{
one or more strings in \code{YYYY-MM-DD} format
}
\description{
Convert one or more dates written in English
to the ISO 8601 \code{YYYY-MM-DD} format
}
\examples{
convert_english_dates_to_iso_8601("25 November 2025") # 2025-11-25
}
