% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_abs_urls.R
\name{extract_abs_urls}
\alias{extract_abs_urls}
\title{Extract an absolute/full URL from a piece of text}
\usage{
extract_abs_urls(text)
}
\arguments{
\item{text}{one line of text}
}
\value{
one single URL without a trailing slash
}
\description{
The text can be messy :-)
}
\examples{
text <- paste0(
  "[Introduction to Bianca: Handling Sensitive Research Data ",
  "<br> _September 18 2026_](https://uppmax.github.io/bianca_workshops/",
  "beginner/intro/){ .md-button .md-button--primary }"
)
extract_abs_url(text) # https://uppmax.github.io/bianca_workshops
}
