% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code_refinery_course_names.R
\name{extract_code_refinery_course_names}
\alias{extract_code_refinery_course_names}
\title{Extract the course names from an Code Refinery courses text}
\usage{
extract_code_refinery_course_names(code_refinery_courses_text)
}
\arguments{
\item{code_refinery_courses_text}{subset of raw HTML from the Code Refinery website.}
}
\description{
Extract the course names from an Code Refinery courses text
}
