% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code_refinery_course_urls.R
\name{extract_code_refinery_course_urls}
\alias{extract_code_refinery_course_urls}
\title{Extract the URL of a course from an Code Refinery courses text}
\usage{
extract_code_refinery_course_urls(code_refinery_courses_text)
}
\arguments{
\item{code_refinery_courses_text}{subset of raw HTML from the Code Refinery website.}
}
\value{
a collection of strings and NAs
}
\description{
Extract the URL of a course from an Code Refinery courses text
}
