% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code_refinery_from_dates.R
\name{extract_code_refinery_from_dates}
\alias{extract_code_refinery_from_dates}
\title{Extract the from dates from an Code Refinery courses text}
\usage{
extract_code_refinery_from_dates(code_refinery_courses_text)
}
\arguments{
\item{code_refinery_courses_text}{subset of raw HTML from the Code Refinery website.}
}
\description{
Extract the from dates from an Code Refinery courses text
}
