% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code_refinery_to_dates.R
\name{extract_code_refinery_to_dates}
\alias{extract_code_refinery_to_dates}
\title{Extract the to dates from an Code Refinery courses text}
\usage{
extract_code_refinery_to_dates(code_refinery_courses_text)
}
\arguments{
\item{code_refinery_courses_text}{subset of raw HTML
from the Code Refinery website.}
}
\description{
Extract the to dates from an Code Refinery courses text
}
