% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_csc_course_names.R
\name{extract_csc_course_names}
\alias{extract_csc_course_names}
\title{Extract the course names from an CSC courses text}
\usage{
extract_csc_course_names(csc_courses_text)
}
\arguments{
\item{csc_courses_text}{subset of raw HTML from the CSC website.}
}
\description{
Extract the course names from an CSC courses text
}
