% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_csc_course_urls.R
\name{extract_csc_course_urls}
\alias{extract_csc_course_urls}
\title{Extract the URL of a course from an CSC courses text}
\usage{
extract_csc_course_urls(csc_courses_text)
}
\arguments{
\item{csc_courses_text}{subset of raw HTML from the CSC website.}
}
\value{
a collection of strings and NAs
}
\description{
Extract the URL of a course from an CSC courses text
}
