% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_csc_from_date.R
\name{extract_csc_from_date}
\alias{extract_csc_from_date}
\title{Extract the from date from an CSC date range}
\usage{
extract_csc_from_date(csc_date_range)
}
\arguments{
\item{csc_date_range}{date range as used at the CSC website.}
}
\value{
from date in ISO 8601 format
}
\description{
Extract the from date from an CSC date range
}
\examples{
extract_csc_from_date("18.-20.11.2025")
extract_csc_from_date("25.11.2025")
extract_csc_from_date("25.-27.11.2025")
extract_csc_from_date("2.-4.12.2025")
}
