% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_csc_to_dates.R
\name{extract_csc_to_dates}
\alias{extract_csc_to_dates}
\title{Extract the to dates from an CSC courses text}
\usage{
extract_csc_to_dates(csc_courses_text)
}
\arguments{
\item{csc_courses_text}{subset of raw HTML from the CSC website.}
}
\description{
Extract the to dates from an CSC courses text
}
