% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_english_from_dates.R
\name{extract_english_from_dates}
\alias{extract_english_from_dates}
\title{Extract the English from dates from English date ranges}
\usage{
extract_english_from_dates(english_date_ranges)
}
\arguments{
\item{english_date_ranges}{English date ranges}
}
\description{
Extract the English from dates from English date ranges
}
