% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_english_ranges.R
\name{extract_english_ranges}
\alias{extract_english_ranges}
\title{Extract the English date ranges from one or more lines of text}
\usage{
extract_english_ranges(text)
}
\arguments{
\item{text}{one or more lines of text}
}
\value{
the English ranges
}
\description{
The text can be messy :-)
}
\examples{
extract_english_ranges("<p>27th Nov 2025</p>") # 27th Nov 2025
}
\seealso{
use \link{extract_english_range} to extract the date range
from one line of text
}
