% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_english_to_date.R
\name{extract_english_to_date}
\alias{extract_english_to_date}
\title{Extract the to date from an English range}
\usage{
extract_english_to_date(english_date_range)
}
\arguments{
\item{english_date_range}{a date range in English}
}
\description{
Extract the to date from an English range
}
\examples{
extract_english_to_date("Nov 07th - Nov 25th, 2022")
}
