% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hpc2n_course_urls.R
\name{extract_hpc2n_course_urls}
\alias{extract_hpc2n_course_urls}
\title{Extract the URL of a course from an HPC2N courses text}
\usage{
extract_hpc2n_course_urls(hpc2n_courses_text)
}
\arguments{
\item{hpc2n_courses_text}{subset of raw HTML from the HPC2N website.}
}
\value{
a collection of strings and NAs
}
\description{
Extract the URL of a course from an HPC2N courses text
}
