% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hpc2n_from_dates.R
\name{extract_hpc2n_from_dates}
\alias{extract_hpc2n_from_dates}
\title{Extract the from dates from an HPC2N courses text}
\usage{
extract_hpc2n_from_dates(hpc2n_courses_text)
}
\arguments{
\item{hpc2n_courses_text}{subset of raw HTML from the HPC2N website.}
}
\description{
Extract the from dates from an HPC2N courses text
}
