% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_infraviz_course_names.R
\name{extract_infraviz_course_names}
\alias{extract_infraviz_course_names}
\title{Extract the course names from an InfraViz courses text}
\usage{
extract_infraviz_course_names(infraviz_courses_text)
}
\arguments{
\item{infraviz_courses_text}{subset of raw HTML from the InfraViz website.}
}
\description{
Extract the course names from an InfraViz courses text
}
