% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_infraviz_course_urls.R
\name{extract_infraviz_course_urls}
\alias{extract_infraviz_course_urls}
\title{Extract the URL of a course from an InfraViz courses text}
\usage{
extract_infraviz_course_urls(infraviz_courses_text)
}
\arguments{
\item{infraviz_courses_text}{subset of raw HTML from the InfraViz website.}
}
\value{
a collection of strings and NAs
}
\description{
Extract the URL of a course from an InfraViz courses text
}
