% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_infraviz_to_dates.R
\name{extract_infraviz_to_dates}
\alias{extract_infraviz_to_dates}
\title{Extract the to dates from an InfraViz courses text}
\usage{
extract_infraviz_to_dates(infraviz_courses_text)
}
\arguments{
\item{infraviz_courses_text}{subset of raw HTML from the InfraViz website.}
}
\description{
Extract the to dates from an InfraViz courses text
}
