% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lunarc_course_names.R
\name{extract_lunarc_course_names}
\alias{extract_lunarc_course_names}
\title{Extract the course names from an HPC2N courses text}
\usage{
extract_lunarc_course_names(lunarc_courses_text)
}
\arguments{
\item{lunarc_courses_text}{subset of raw HTML from the HPC2N website.}
}
\description{
Extract the course names from an HPC2N courses text
}
