% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lunarc_course_urls.R
\name{extract_lunarc_course_urls}
\alias{extract_lunarc_course_urls}
\title{Extract the URL of a course from an LUNARC courses text}
\usage{
extract_lunarc_course_urls(
  lunarc_courses_text,
  lunarc_home_url = "https://www.lunarc.lu.se/"
)
}
\arguments{
\item{lunarc_courses_text}{subset of raw HTML from the LUNARC website.}

\item{lunarc_home_url}{The URL for the LUNARC home}
}
\value{
a collection of strings and NAs
}
\description{
Extract the URL of a course from an LUNARC courses text
}
