% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lunarc_from_dates.R,
%   R/extract_lunarc_to_dates.R
\name{extract_lunarc_from_dates}
\alias{extract_lunarc_from_dates}
\title{Extract the from dates from a LUNARC courses text}
\usage{
extract_lunarc_from_dates(lunarc_courses_text)

extract_lunarc_from_dates(lunarc_courses_text)
}
\arguments{
\item{lunarc_courses_text}{subset of raw HTML from the LUNARC website.}
}
\description{
Extract the from dates from a LUNARC courses text

Extract the from dates from a LUNARC courses text
}
