% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nsc_course_names.R
\name{extract_nsc_course_names}
\alias{extract_nsc_course_names}
\title{Extract the course names from an NSC courses text}
\usage{
extract_nsc_course_names(nsc_courses_text)
}
\arguments{
\item{nsc_courses_text}{subset of raw HTML from the NSC website.}
}
\description{
Extract the course names from an NSC courses text
}
