% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nsc_course_urls.R
\name{extract_nsc_course_urls}
\alias{extract_nsc_course_urls}
\title{Extract the URL of a course from an NSC courses text}
\usage{
extract_nsc_course_urls(
  nsc_courses_text,
  nsc_home_url = scoreto::get_provider_home_url("NSC")
)
}
\arguments{
\item{nsc_courses_text}{subset of raw HTML from the NSC website.}

\item{nsc_home_url}{The URL for the NSC home}
}
\value{
a collection of strings and NAs
}
\description{
Extract the URL of a course from an NSC courses text
}
