% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nsc_to_dates.R
\name{extract_nsc_to_dates}
\alias{extract_nsc_to_dates}
\title{Extract the to dates from an NSC courses text}
\usage{
extract_nsc_to_dates(nsc_courses_text)
}
\arguments{
\item{nsc_courses_text}{subset of raw HTML from the NSC website.}
}
\description{
Extract the to dates from an NSC courses text
}
