% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pdc_course_names.R
\name{extract_pdc_course_names}
\alias{extract_pdc_course_names}
\title{Extract the course names from an PDC courses text}
\usage{
extract_pdc_course_names(pdc_courses_text)
}
\arguments{
\item{pdc_courses_text}{subset of raw HTML from the PDC website.}
}
\description{
Extract the course names from an PDC courses text
}
