% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pdc_to_dates.R
\name{extract_pdc_to_dates}
\alias{extract_pdc_to_dates}
\title{Extract the to dates from an PDC courses text}
\usage{
extract_pdc_to_dates(pdc_courses_text)
}
\arguments{
\item{pdc_courses_text}{subset of raw HTML from the PDC website.}
}
\description{
Extract the to dates from an PDC courses text
}
