% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rel_url.R
\name{extract_rel_url}
\alias{extract_rel_url}
\title{Extract a relative URL from a piece of text}
\usage{
extract_rel_url(text)
}
\arguments{
\item{text}{one line of text}
}
\value{
one single relative URL without a trailing slash
}
\description{
The text can be messy :-)
}
\examples{
text <- paste0(
  "<p><a class=\"md-button md-button--primary\" ",
  "href=\"../uppmax_intro_course/\">Introduction to Linux and ",
  "UPPMAX <br> <em>August 24-26 2026</em></a></p>"
)
extract_rel_url(text) # ../uppmax_intro_course
}
