% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rel_urls.R
\name{extract_rel_urls}
\alias{extract_rel_urls}
\title{Extract one or more relative URL from a piece of text.}
\usage{
extract_rel_urls(text)
}
\arguments{
\item{text}{one line of text}
}
\value{
per line, one single relative URL (without a trailing slash) or NA
}
\description{
Assumes that there is no or one relative URL per line.
}
\details{
The text can be messy :-)
}
\examples{
text <- paste0(
  "<p><a class=\"md-button md-button--primary\" ",
  "href=\"../uppmax_intro_course/\">Introduction to Linux and ",
  "UPPMAX <br> <em>August 24-26 2026</em></a></p>"
)
extract_rel_urls(text) # ../uppmax_intro_course
}
