% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sll_to_date.R
\name{extract_sll_to_date}
\alias{extract_sll_to_date}
\title{Extract the 'to' date}
\usage{
extract_sll_to_date(sll_date_range)
}
\arguments{
\item{sll_date_range}{a SciLifeLab date range}
}
\value{
a string in \code{YYYY-MM-DD} format
}
\description{
Extract the 'to' date
}
