% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_slubi_course_urls.R
\name{extract_slubi_course_urls}
\alias{extract_slubi_course_urls}
\title{Extract the URL of a course from an SLUBI courses text}
\usage{
extract_slubi_course_urls(slubi_courses_text)
}
\arguments{
\item{slubi_courses_text}{subset of raw HTML from the SLUBI website.}
}
\value{
a collection of strings and NAs
}
\description{
Extract the URL of a course from an SLUBI courses text
}
