% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_slubi_from_date.R
\name{extract_slubi_from_date}
\alias{extract_slubi_from_date}
\title{Extract the from date from an SLUBI courses text}
\usage{
extract_slubi_from_date(english_date_range)
}
\arguments{
\item{english_date_range}{a data range in English}
}
\description{
Extract the from date from an SLUBI courses text
}
\examples{
extract_slubi_from_date("Nov 07th - Nov 25th, 2022")

}
