% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_slubi_to_date.R
\name{extract_slubi_to_date}
\alias{extract_slubi_to_date}
\title{Extract the to date from an SLUBI courses text}
\usage{
extract_slubi_to_date(english_to_date)
}
\arguments{
\item{english_to_date}{one date from the SLUBI website.}
}
\description{
Extract the to date from an SLUBI courses text
}
