% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_slubi_to_dates.R
\name{extract_slubi_to_dates}
\alias{extract_slubi_to_dates}
\title{Extract the to dates from an SLUBI courses text}
\usage{
extract_slubi_to_dates(english_to_dates)
}
\arguments{
\item{english_dates}{subset of raw HTML from the SLUBI website.}
}
\description{
Extract the to dates from an SLUBI courses text
}
