% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_uppmax_from_dates.R
\name{extract_uppmax_from_dates}
\alias{extract_uppmax_from_dates}
\title{Extract the 'from' date}
\usage{
extract_uppmax_from_dates(uppmax_date_ranges)
}
\arguments{
\item{uppmax_date_range}{a SciLifeLab date range}
}
\value{
a string in \code{YYYY-MM-DD} format
}
\description{
Extract the 'from' date
}
\examples{
extract_uppmax_from_date("Aug 25-27") # 2025-08-25
extract_uppmax_from_date("Oct 14") # 2025-10-14
extract_uppmax_from_date("Oct 23-24 + 27-29") # 2025-10-23
}
