% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_uppmax_to_date.R
\name{extract_uppmax_to_date}
\alias{extract_uppmax_to_date}
\title{Extract the 'to' date}
\usage{
extract_uppmax_to_date(uppmax_date_range)
}
\arguments{
\item{uppmax_date_range}{an UPPMAX date range}
}
\value{
a string in \code{YYYY-MM-DD} format
}
\description{
Extract the 'to' date
}
\examples{
extract_uppmax_to_date("Aug 25-27") # 2025-08-27
extract_uppmax_to_date("Oct 14") # 2025-10-14
extract_uppmax_to_date("Oct 23-24 + 27-29") # 2025-10-29
}
