% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_uppmax_to_dates.R
\name{extract_uppmax_to_dates}
\alias{extract_uppmax_to_dates}
\title{Extract the 'to' date from one or more UPPMAX dates}
\usage{
extract_uppmax_to_dates(uppmax_date_ranges)
}
\arguments{
\item{uppmax_date_range}{a UPPMAX date range}
}
\value{
a string in \code{YYYY-MM-DD} format
}
\description{
Extract the 'to' date from one or more UPPMAX dates
}
\examples{
extract_uppmax_to_dates("Aug 25-27") # 2025-08-27
extract_uppmax_to_dates("Oct 14") # 2025-10-14
extract_uppmax_to_dates("Oct 23-24 + 27-29") # 2025-10-29
}
