% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_providers.R
\name{get_all_providers}
\alias{get_all_providers}
\title{Get the names of all the providers of computational resources}
\usage{
get_all_providers()
}
\value{
the names of all the providers in alphabetic order
}
\description{
Get the names of all the providers of computational resources
}
\examples{
get_all_providers()
}
