% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_providers_info.R
\name{get_all_providers_info}
\alias{get_all_providers_info}
\title{Get a table with all of the providers' information}
\usage{
get_all_providers_info()
}
\value{
a table with all providers' information,
}
\description{
Get a table with all of the providers' information
}
