% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_any_month_regex.R
\name{get_any_month_regex}
\alias{get_any_month_regex}
\title{Get a regular expression that matches any month}
\usage{
get_any_month_regex()
}
\value{
a regular expression for detecting a month
}
\description{
Get a regular expression that matches any month
}
\note{
I did not expect that I would need to write this function
}
