% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_c3se_courses.R
\name{get_c3se_courses}
\alias{get_c3se_courses}
\title{Get the C3SE courses}
\usage{
get_c3se_courses(html_text = scoreto::get_c3se_html())
}
\arguments{
\item{html_text}{HTML text to parse, as can be obtained by
\link{get_c3se_html}}
}
\value{
a table with all C3SE courses.
}
\description{
Get the C3SE courses
}
