% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_c3se_html.R
\name{get_c3se_html}
\alias{get_c3se_html}
\title{Get the C3SE courses as HTML}
\usage{
get_c3se_html(c3se_training_url = scoreto::get_provider_courses_url("C3SE"))
}
\arguments{
\item{c3se_training_url}{the URL of the C3SE training webpage}
}
\value{
the HTML text of the C3SE courses.
}
\description{
Get the C3SE courses as HTML
}
\seealso{
use \link{get_c3se_courses} to get the C3SE courses as a table.
}
