% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdhu_courses.R
\name{get_cdhu_courses}
\alias{get_cdhu_courses}
\title{Get the CDHU courses}
\usage{
get_cdhu_courses()
}
\value{
a table with all CDHU courses.
}
\description{
Get the CDHU courses
}
\note{
The text that CDHU provides is too complex to parse.
If the text is written for humans, then let humans decipher it.
Until the text at CDHU is made machine-readable
in any reasonable way, the courses of CDHU will not be shown
}
