% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_to_center.R
\name{get_cluster_to_center}
\alias{get_cluster_to_center}
\title{Get a table of all the HPC clusters and the center(s) these
are connected to.}
\usage{
get_cluster_to_center()
}
\value{
a table with all HPC clusters and the center(s) these
are connected to, sorted alphabetically by HPC cluster name
}
\description{
Get a table of all the HPC clusters and the center(s) these
are connected to.
}
