% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code_refinery_courses.R
\name{get_code_refinery_courses}
\alias{get_code_refinery_courses}
\title{Get the Code Refinery courses}
\usage{
get_code_refinery_courses(html_text = scoreto::get_code_refinery_html())
}
\value{
a table with all Code Refinery courses.
}
\description{
Get the Code Refinery courses
}
