% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code_refinery_courses_url.R
\name{get_code_refinery_courses_url}
\alias{get_code_refinery_courses_url}
\title{Get the URL to the Code Refinery courses page}
\usage{
get_code_refinery_courses_url()
}
\value{
the URL to the Code Refinery courses page
}
\description{
Get the URL to the Code Refinery courses page
}
