% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code_refinery_html.R
\name{get_code_refinery_html}
\alias{get_code_refinery_html}
\title{Get the Code Refinery courses website as HTML}
\usage{
get_code_refinery_html()
}
\value{
HTML text
}
\description{
Get the Code Refinery courses website as HTML
}
\seealso{
use \link{get_test_code_refinery_html} to get similar HTML text to be
used in testing
}
