% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code_refinery_test_courses_text.R
\name{get_code_refinery_test_courses_text}
\alias{get_code_refinery_test_courses_text}
\title{Get a testing text for the Code Refinery courses}
\usage{
get_code_refinery_test_courses_text()
}
\description{
Get a testing text for the Code Refinery courses
}
