% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compute.R
\name{get_compute}
\alias{get_compute}
\title{Get a table with all of the compute resources}
\usage{
get_compute()
}
\value{
a table with all compute courses,
sorted alphabetically by resource name
}
\description{
This table is displayed at
\url{https://nbisweden.github.io/SCoRe_user_doc/compute/}.
}
\seealso{
use \link{get_compute_as_markdown} to get this
table as Markdown text.
}
