% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compute_as_markdown.R
\name{get_compute_as_markdown}
\alias{get_compute_as_markdown}
\title{Get all the compute resources, as a Markdown table}
\usage{
get_compute_as_markdown(t = scoreto::get_compute())
}
\arguments{
\item{t}{a table, as returned by \link{get_compute}}
}
\value{
a Markdown table of all compute resources, as text
}
\description{
This table is displayed at
\url{https://nbisweden.github.io/SCoRe_user_doc/compute/}.
}
\seealso{
use \link{get_compute} to get this
table as a table.
}
