% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_courses_as_markdown.R
\name{get_courses_as_markdown}
\alias{get_courses_as_markdown}
\title{Get all the courses, as a Markdown table}
\usage{
get_courses_as_markdown()
}
\value{
a Markdown table  of all courses, as text
}
\description{
Get all the courses, as a Markdown table
}
