% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csc_courses.R
\name{get_csc_courses}
\alias{get_csc_courses}
\title{Get the CSC courses}
\usage{
get_csc_courses(html_text = scoreto::get_csc_html())
}
\arguments{
\item{html_text}{HTML text to parse, as can be obtained by
\link{get_csc_html} or \link{get_test_csc_html}}
}
\value{
a table with all CSC courses.
}
\description{
Get the CSC courses
}
