% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csc_courses_url.R
\name{get_csc_courses_url}
\alias{get_csc_courses_url}
\title{Get the URL to the CSC courses page}
\usage{
get_csc_courses_url()
}
\value{
the URL to the CSC courses page
}
\description{
Get the URL to the CSC courses page
}
