% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csc_html.R
\name{get_csc_html}
\alias{get_csc_html}
\title{Get the CSC courses website as HTML}
\usage{
get_csc_html()
}
\value{
HTML text
}
\description{
Get the CSC courses website as HTML
}
\seealso{
use \link{get_test_csc_html} to get similar HTML text to be
used in testing
}
