% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csc_test_courses_text.R
\name{get_csc_test_courses_text}
\alias{get_csc_test_courses_text}
\title{Get a testing text for CSC courses}
\usage{
get_csc_test_courses_text()
}
\description{
Get a testing text for CSC courses
}
