% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enccs_courses.R
\name{get_enccs_courses}
\alias{get_enccs_courses}
\title{Get the ENCCS courses}
\usage{
get_enccs_courses(html_text = scoreto::get_enccs_html())
}
\value{
a table with all ENCCS courses.
}
\description{
Get the ENCCS courses
}
