% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enccs_courses_url.R
\name{get_enccs_courses_url}
\alias{get_enccs_courses_url}
\title{Get the URL to the ENCCS courses page}
\usage{
get_enccs_courses_url()
}
\value{
the URL to the ENCCS courses page
}
\description{
Get the URL to the ENCCS courses page
}
\examples{
get_enccs_courses_url()
}
