% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enccs_html.R
\name{get_enccs_html}
\alias{get_enccs_html}
\title{Get the ENCCS courses website as HTML}
\usage{
get_enccs_html()
}
\value{
HTML text
}
\description{
Get the ENCCS courses website as HTML
}
\seealso{
use \link{get_test_enccs_html} to get similar HTML text to be
used in testing
}
