% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hpc2n_courses_url.R
\name{get_hpc2n_courses_url}
\alias{get_hpc2n_courses_url}
\title{Get the URL to the HPC2N courses page}
\usage{
get_hpc2n_courses_url()
}
\value{
the URL to the HPC2N courses page
}
\description{
Get the URL to the HPC2N courses page
}
