% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infraviz_course_info.R
\name{get_infraviz_course_info}
\alias{get_infraviz_course_info}
\title{Extract the course info from an InfraViz page}
\usage{
get_infraviz_course_info(course_page_url)
}
\arguments{
\item{course_page_url}{URL to the course page}
}
\value{
a \link[tibble]{tibble}
with column names \code{date_from} and \code{date_to}
}
\description{
Extract the course info from an InfraViz page
}
