% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infraviz_courses_infos.R
\name{get_infraviz_courses_infos}
\alias{get_infraviz_courses_infos}
\title{Extract the course info from an InfraViz page}
\usage{
get_infraviz_courses_infos(course_pages_urls)
}
\arguments{
\item{course_pages_urls}{one or more URLs to the course page(s)}
}
\value{
a \link{tibble::tibble}
with column names \code{date_from} and \code{date_to}
}
\description{
Extract the course info from an InfraViz page
}
\seealso{
use \link{get_infraviz_course_info} for one course page URL
}
