% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infraviz_courses_url.R
\name{get_infraviz_courses_url}
\alias{get_infraviz_courses_url}
\title{Get the URL to the InfraViz courses page}
\usage{
get_infraviz_courses_url()
}
\value{
the URL to the InfraViz courses page
}
\description{
Get the URL to the InfraViz courses page
}
\examples{
get_infraviz_courses_url()
}
