% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infraviz_html.R
\name{get_infraviz_html}
\alias{get_infraviz_html}
\title{Get the InfraViz courses website as HTML}
\usage{
get_infraviz_html()
}
\value{
HTML text
}
\description{
Get the InfraViz courses website as HTML
}
\seealso{
use \link{get_test_infraviz_html} to get similar HTML text to be
used in testing
}
