% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infraviz_test_courses_text.R
\name{get_infraviz_test_courses_text}
\alias{get_infraviz_test_courses_text}
\title{Get a testing text for InfraViz courses}
\usage{
get_infraviz_test_courses_text()
}
\description{
Get a testing text for InfraViz courses
}
