% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logo_path.R
\name{get_logo_path}
\alias{get_logo_path}
\title{Get the path to the logo of a provider of computational resources}
\usage{
get_logo_path(provider_name)
}
\arguments{
\item{provider_name}{the name of a provider of computational resources}
}
\value{
the path to the logo
}
\description{
Get the path to the logo of a provider of computational resources
}
\seealso{
Use \link{get_all_providers} to get a list of all providers
}
