% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logo_paths.R
\name{get_logo_paths}
\alias{get_logo_paths}
\title{Get the paths to the logos of a providers of computational resources}
\usage{
get_logo_paths(provider_names)
}
\arguments{
\item{provider_names}{the names of a provider of computational resources}
}
\value{
the paths to the logos
}
\description{
Get the paths to the logos of a providers of computational resources
}
