% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lunarc_courses.R
\name{get_lunarc_courses}
\alias{get_lunarc_courses}
\title{Get the LUNARC courses}
\usage{
get_lunarc_courses(html_text = get_lunarc_html())
}
\arguments{
\item{html_text}{HTML text to parse, as can be obtained by
\link{get_lunarc_html} or \link{get_test_lunarc_html}}
}
\value{
a table with all LUNARC courses.
}
\description{
Get the LUNARC courses
}
