% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lunarc_courses_url.R
\name{get_lunarc_courses_url}
\alias{get_lunarc_courses_url}
\title{Get the URL to the LUNARC courses page}
\usage{
get_lunarc_courses_url()
}
\value{
the URL to the LUNARC courses page
}
\description{
Get the URL to the LUNARC courses page
}
