% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lunarc_html.R
\name{get_lunarc_html}
\alias{get_lunarc_html}
\title{Get the LUNARC courses website as HTML}
\usage{
get_lunarc_html()
}
\value{
HTML text
}
\description{
Get the LUNARC courses website as HTML
}
\seealso{
use \link{get_test_lunarc_html} to get similar HTML text to be
used in testing
}
