% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mimer_courses.R
\name{get_mimer_courses}
\alias{get_mimer_courses}
\title{Get the Mimer courses}
\usage{
get_mimer_courses(html_text = scoreto::get_mimer_html())
}
\arguments{
\item{html_text}{HTML text to parse, as can be obtained by
\link{get_mimer_html} or \link{get_test_mimer_html}}
}
\value{
a table with all Mimer courses, where
the table will pass the test of
\link{is_correctly_formatted_courses_table}
}
\description{
Get the Mimer courses
}
