% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mimer_courses_url.R
\name{get_mimer_courses_url}
\alias{get_mimer_courses_url}
\title{Get the URL to the Mimer courses page}
\usage{
get_mimer_courses_url()
}
\value{
the URL to the Mimer courses page
}
\description{
Get the URL to the Mimer courses page
}
