% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mimer_html.R
\name{get_mimer_html}
\alias{get_mimer_html}
\title{Get the Mimer events website as HTML}
\usage{
get_mimer_html()
}
\value{
HTML text
}
\description{
Get the Mimer events website as HTML
}
\seealso{
use \link{get_test_mimer_html} to get similar HTML text to be
used in testing
}
