% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naiss_courses_url.R
\name{get_naiss_courses_url}
\alias{get_naiss_courses_url}
\title{Get the URL to the NAISS courses page}
\usage{
get_naiss_courses_url()
}
\value{
the URL to the NAISS courses page
}
\description{
Get the URL to the NAISS courses page
}
