% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naiss_html.R
\name{get_naiss_html}
\alias{get_naiss_html}
\title{Get the NAISS courses website as HTML}
\usage{
get_naiss_html()
}
\value{
HTML text
}
\description{
Get the NAISS courses website as HTML
}
\seealso{
use \link{get_test_naiss_html} to get similar HTML text to be
used in testing
}
