% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nsc_courses.R
\name{get_nsc_courses}
\alias{get_nsc_courses}
\title{Get the NSC courses}
\usage{
get_nsc_courses(html_text = scoreto::get_nsc_html())
}
\arguments{
\item{html_text}{HTML text to parse, as can be obtained by
\link{get_nsc_html} or \link{get_test_nsc_html}}
}
\value{
a table with all NSC courses, where
the table will pass the test of
\link{is_correctly_formatted_courses_table}
}
\description{
Get the NSC courses
}
\examples{
html_text <- readr::read_lines(get_scoreto_path("nsc.html"))
get_nsc_courses(html_text)
}
