% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nsc_html.R
\name{get_nsc_html}
\alias{get_nsc_html}
\title{Get the NSC courses as HTML}
\usage{
get_nsc_html(nsc_training_url = "https://www.nsc.liu.se/support/Events/")
}
\arguments{
\item{nsc_training_url}{the URL of the NSC training webpage}
}
\value{
the HTML text of the NSC courses.
}
\description{
Get the NSC courses as HTML
}
\examples{
get_nsc_html()
}
\seealso{
use \link{get_nsc_courses} to get the NSC courses as a table
}
