% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nsc_test_courses_text.R
\name{get_nsc_test_courses_text}
\alias{get_nsc_test_courses_text}
\title{Get a testing text for the NSC courses}
\usage{
get_nsc_test_courses_text()
}
\description{
Get a testing text for the NSC courses
}
