% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nsc_training_url.R
\name{get_nsc_training_url}
\alias{get_nsc_training_url}
\title{Get the URL to the NSC courses}
\usage{
get_nsc_training_url()
}
\value{
the URL to the NSC courses
}
\description{
Get the URL to the NSC courses
}
\examples{
get_nsc_training_url()
}
