% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_md_table_divider.R
\name{get_optimal_markdown_divider}
\alias{get_optimal_markdown_divider}
\title{Get the optimal divider for a Markdown table}
\usage{
get_optimal_markdown_divider(text)
}
\arguments{
\item{text}{a text that is a markdown table,
as returned by \code{knitr::kable}}
}
\description{
Get the optimal divider for a Markdown table
}
