% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oscu_events.R
\name{get_oscu_events}
\alias{get_oscu_events}
\title{Get the OSCU events}
\usage{
get_oscu_events(html_text = scoreto::get_oscu_html())
}
\arguments{
\item{html_text}{HTML text to parse, as can be obtained by
\link{get_oscu_html} or \link{get_test_oscu_html}}
}
\value{
a table with all OSCU events, where
the table will pass the test of
\link{is_correctly_formatted_courses_table}
}
\description{
Get the OSCU events
}
