% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oscu_events_url.R
\name{get_oscu_events_url}
\alias{get_oscu_events_url}
\title{Get the URL to the OSCU events page}
\usage{
get_oscu_events_url()
}
\value{
the URL to the OSCU events page
}
\description{
Get the URL to the OSCU events page
}
