% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oscu_html.R
\name{get_oscu_html}
\alias{get_oscu_html}
\title{Get the OSCU events website as HTML}
\usage{
get_oscu_html()
}
\value{
HTML text
}
\description{
Get the OSCU events website as HTML
}
\seealso{
use \link{get_test_oscu_html} to get similar HTML text to be
used in testing
}
