% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oscu_test_events_text.R
\name{get_oscu_test_events_text}
\alias{get_oscu_test_events_text}
\title{Get a testing text for OSCU events}
\usage{
get_oscu_test_events_text()
}
\value{
a text to extract the dates from
}
\description{
This text is a patchwork of entries from the OSCU website
collected over time.
}
