% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_other.R
\name{get_other}
\alias{get_other}
\title{Get a table with all of the other resources}
\usage{
get_other()
}
\value{
a table with all other resources,
sorted alphabetically by other name
}
\description{
This table is displayed at
\url{https://nbisweden.github.io/SCoRe_user_doc/other/}.
}
\seealso{
use \link{get_other_as_markdown} to get this
table as Markdown text.
}
