% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_other_as_markdown.R
\name{get_other_as_markdown}
\alias{get_other_as_markdown}
\title{Get all the other resources, as a Markdown table}
\usage{
get_other_as_markdown(t = get_other())
}
\arguments{
\item{t}{a table, as returned by \link{get_other}}
}
\value{
a Markdown table of all other resources, as text
}
\description{
This table is displayed at
\url{https://nbisweden.github.io/SCoRe_user_doc/other/}.
}
\seealso{
use \link{get_other} to get this
table as a table.
}
