% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_other_to_center.R
\name{get_other_to_center}
\alias{get_other_to_center}
\title{Get a table of all the others and the center(s) these
are connected to.}
\usage{
get_other_to_center()
}
\value{
a table with all others and the center(s) these
are connected to, in tidy data format,
sorted alphabetically by (1) other name,
(2) center name
}
\description{
Get a table of all the others and the center(s) these
are connected to.
}
\examples{
get_other_to_center()
}
