% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdc_html.R
\name{get_pdc_html}
\alias{get_pdc_html}
\title{Get the PDC courses as HTML}
\usage{
get_pdc_html(pdc_training_url = scoreto::get_provider_courses_url("PDC"))
}
\arguments{
\item{pdc_training_url}{the URL of the PDC training webpage}
}
\value{
the HTML text of the PDC courses.
}
\description{
Get the PDC courses as HTML
}
\seealso{
use \link{get_pdc_courses} to get the PDC courses as a table.
}
