% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdc_test_courses_text.R
\name{get_pdc_test_courses_text}
\alias{get_pdc_test_courses_text}
\title{Get a testing text for the PDC courses}
\usage{
get_pdc_test_courses_text()
}
\description{
Get a testing text for the PDC courses
}
