% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provider_courses_url.R
\name{get_provider_courses_url}
\alias{get_provider_courses_url}
\title{Get a URL to the courses page of a provider}
\usage{
get_provider_courses_url(provider_name)
}
\arguments{
\item{provider_name}{name of the provider}
}
\description{
Get a URL to the courses page of a provider
}
\examples{
get_provider_courses_url("UPPMAX")
}
\seealso{
\link{get_provider_home_url}
}
