% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provider_home_url.R
\name{get_provider_home_url}
\alias{get_provider_home_url}
\title{Get a URL to the homepage of a provider}
\usage{
get_provider_home_url(provider_name)
}
\arguments{
\item{provider_name}{name of the provider}
}
\description{
Get a URL to the homepage of a provider
}
\examples{
get_provider_home_url("UPPMAX")
}
\seealso{
\itemize{
\item \link{get_providers_home_urls}: get the URLs to the homepages
of multiple providers
\item \link{get_provider_courses_url}: for the courses page of the provider
}
}
