% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provider_logo_as_markdown.R
\name{get_provider_logo_as_markdown}
\alias{get_provider_logo_as_markdown}
\title{Get the logo of the provider as Markdown}
\usage{
get_provider_logo_as_markdown(provider_name)
}
\value{
the logo of the providers as Markdown
}
\description{
Get the logo of the provider as Markdown
}
\examples{
get_provider_logo_as_markdown("UPPMAX")
}
\seealso{
use \link{get_all_provider_logos_as_markdown} to get the
Markdown of all these
}
