% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_providers_home_urls.R
\name{get_providers_home_urls}
\alias{get_providers_home_urls}
\title{get the URLs to the homepages of multiple providers}
\usage{
get_providers_home_urls(provider_names)
}
\description{
get the URLs to the homepages of multiple providers
}
\examples{
get_providers_home_urls(c("NAISS", "UPPMAX"))
}
\seealso{
\itemize{
\item \link{get_provider_home_url}: get the URLs to one homepage
of a provider
}
}
