% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scoreto_path.R
\name{get_scoreto_path}
\alias{get_scoreto_path}
\title{Get the full path of a file in the \code{inst/extdata} folder}
\usage{
get_scoreto_path(filename)
}
\arguments{
\item{filename}{the file's name, without the path description}
}
\value{
the full path of the filename
}
\description{
Get the full path of a file in the \code{inst/extdata} folder
}
\examples{
get_scoreto_path("compute.csv")
}
\author{
Richèl J.C. Bilderbeek
}
