% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sll_courses.R
\name{get_sll_courses}
\alias{get_sll_courses}
\title{Get the SciLifeLab courses}
\usage{
get_sll_courses()
}
\value{
a table with all SciLifeLab courses.

This table is a \link[tibble]{tibble} with these column names:
\itemize{
\item \code{source}: the URL where this course was found
\item \code{name}: the name of the course
\item \code{url}: the URL of the course
}
}
\description{
Get the SciLifeLab courses
}
