% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sll_courses_url.R
\name{get_sll_courses_url}
\alias{get_sll_courses_url}
\title{Get the URL to the SciLifeLab courses page}
\usage{
get_sll_courses_url()
}
\value{
the URL to the SciLifeLab courses page
}
\description{
Get the URL to the SciLifeLab courses page
}
