% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sll_html.R
\name{get_sll_html}
\alias{get_sll_html}
\title{Get the SciLifeLab courses website as HTML}
\usage{
get_sll_html()
}
\value{
HTML text
}
\description{
Get the SciLifeLab courses website as HTML
}
\seealso{
use \link{get_test_sll_html} to get similar HTML text to be
used in testing
}
