% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slubi_courses_url.R
\name{get_slubi_courses_url}
\alias{get_slubi_courses_url}
\title{Get the URL to the SLUBI courses page}
\usage{
get_slubi_courses_url()
}
\value{
the URL to the SLUBI courses page
}
\description{
Get the URL to the SLUBI courses page
}
