% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slubi_html.R
\name{get_slubi_html}
\alias{get_slubi_html}
\title{Get the HPC2n courses website as HTML}
\usage{
get_slubi_html()
}
\value{
HTML text
}
\description{
Get the HPC2n courses website as HTML
}
\seealso{
use \link{get_test_slubi_html} to get similar HTML text to be
used in testing
}
