% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slubi_test_courses_text.R
\name{get_slubi_test_courses_text}
\alias{get_slubi_test_courses_text}
\title{Get a testing text for SLUBI courses}
\usage{
get_slubi_test_courses_text()
}
\value{
a text to extract the dates from
}
\description{
This text is a patchwork of entries from the SLUBI website
collected over time.
}
