% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storage.R
\name{get_storage}
\alias{get_storage}
\title{Get a table with all of the storage resources}
\usage{
get_storage(check_file = FALSE)
}
\arguments{
\item{check_file}{Check the file before reading it.
This is used for debugging and can be assumed to not be needed.}
}
\value{
a table with all storage resources,
sorted alphabetically by storage name
}
\description{
This table is displayed at
\url{https://nbisweden.github.io/SCoRe_user_doc/storage/}.
}
\seealso{
use \link{get_storage_as_markdown} to get this
table as Markdown text.
}
