% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storage_as_markdown.R
\name{get_storage_as_markdown}
\alias{get_storage_as_markdown}
\title{Get all the storage resources, as a Markdown table}
\usage{
get_storage_as_markdown(t = get_storage())
}
\arguments{
\item{t}{a table, as returned by \link{get_storage}}
}
\value{
a Markdown table of all storage resources, as text
}
\description{
This table is displayed at
\url{https://nbisweden.github.io/SCoRe_user_doc/storage/}.
}
\seealso{
use \link{get_storage} to get this
table as a table.
}
