% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_code_refinery_html.R
\name{get_test_code_refinery_html}
\alias{get_test_code_refinery_html}
\title{Get testing HTML of the Code Refinery courses website}
\usage{
get_test_code_refinery_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the Code Refinery courses website
}
\seealso{
use \link{get_code_refinery_html} to get HTML text read from the
Code Refinery courses website
}
