% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_courses_table.R
\name{get_test_courses_table}
\alias{get_test_courses_table}
\title{Get a table of courses, to be used in testing.}
\usage{
get_test_courses_table()
}
\value{
a table of all courses, to be used in testing.
The second row contains a course that has an invalied \code{course_url}.
}
\description{
Get a table of courses, to be used in testing,
similar to the one returned by (the much slower) \link{get_courses}
}
