% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_csc_html.R
\name{get_test_csc_html}
\alias{get_test_csc_html}
\title{Get testing HTML of the CSC courses website}
\usage{
get_test_csc_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the CSC courses website
}
\seealso{
use \link{get_csc_html} to get HTML text read from the
CSC courses website
}
