% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_enccs_html.R
\name{get_test_enccs_html}
\alias{get_test_enccs_html}
\title{Get testing HTML of the ENCCS courses website}
\usage{
get_test_enccs_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the ENCCS courses website
}
\seealso{
use \link{get_enccs_html} to get HTML text read from the
ENCCS courses website
}
