% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_infraviz_html.R
\name{get_test_infraviz_html}
\alias{get_test_infraviz_html}
\title{Get testing HTML of the InfraViz courses website}
\usage{
get_test_infraviz_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the InfraViz courses website
}
\seealso{
use \link{get_infraviz_html} to get HTML text read from the
InfraViz courses website
}
