% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_lunarc_html.R
\name{get_test_lunarc_html}
\alias{get_test_lunarc_html}
\title{Get testing HTML of the LUNARC courses website}
\usage{
get_test_lunarc_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the LUNARC courses website
}
\seealso{
use \link{get_lunarc_html} to get HTML text read from the
LUNARC courses website
}
