% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_mimer_html.R
\name{get_test_mimer_html}
\alias{get_test_mimer_html}
\title{Get testing HTML of the Mimer courses website}
\usage{
get_test_mimer_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the Mimer courses website
}
\seealso{
use \link{get_mimer_html} to get HTML text read from the
Mimer courses website
}
