% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_naiss_html.R
\name{get_test_naiss_html}
\alias{get_test_naiss_html}
\title{Get testing HTML of the NAISS courses website}
\usage{
get_test_naiss_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the NAISS courses website
}
\seealso{
use \link{get_naiss_html} to get HTML text read from the
NAISS courses website
}
