% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_oscu_html.R
\name{get_test_oscu_html}
\alias{get_test_oscu_html}
\title{Get testing HTML of the OSCU courses website}
\usage{
get_test_oscu_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the OSCU courses website
}
\seealso{
use \link{get_oscu_html} to get HTML text read from the
OSCU courses website
}
