% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_pdc_html.R
\name{get_test_pdc_html}
\alias{get_test_pdc_html}
\title{Get testing HTML of the PDC courses website}
\usage{
get_test_pdc_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the PDC courses website
}
\seealso{
use \link{get_pdc_html} to get HTML text read from the
PDC courses website
}
