% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_sll_html.R
\name{get_test_sll_html}
\alias{get_test_sll_html}
\title{Get testing HTML of the SciLifeLab courses website}
\usage{
get_test_sll_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the SciLifeLab courses website
}
\seealso{
use \link{get_sll_html} to get HTML text read from the
SLUBI courses website
}
