% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_slubi_html.R
\name{get_test_slubi_html}
\alias{get_test_slubi_html}
\title{Get testing HTML of the HPC2n courses website}
\usage{
get_test_slubi_html()
}
\value{
HTML text
}
\description{
Get testing HTML of the HPC2n courses website
}
\seealso{
use \link{get_slubi_html} to get HTML text read from the
SLUBI courses website
}
