% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_url.R
\name{get_training_url}
\alias{get_training_url}
\title{Get the URL to the courses of a provider}
\usage{
get_training_url(course_provider)
}
\arguments{
\item{course_provider}{name of the course provider}
}
\value{
the URL to the courses of a provider
}
\description{
Get the URL to the courses of a provider
}
\examples{
get_training_url("UPPMAX")
}
