% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uppmax_courses.R
\name{get_uppmax_courses}
\alias{get_uppmax_courses}
\title{Get the NAISS courses}
\usage{
get_uppmax_courses(html_text = scoreto::get_uppmax_html())
}
\arguments{
\item{html_text}{HTML text to parse, as can be obtained by
\link{get_uppmax_html}}
}
\value{
a table with all UPPMAX courses, where
the table will pass the test of
\link{is_correctly_formatted_courses_table}
}
\description{
Get the NAISS courses
}
