% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_web_host.R
\name{get_web_host}
\alias{get_web_host}
\title{Get a table with all of the web_host resources}
\usage{
get_web_host()
}
\value{
a table with all web_host resources,
sorted alphabetically by web_host name
}
\description{
This table is displayed at
\url{https://nbisweden.github.io/SCoRe_user_doc/web_host/}.
}
\seealso{
use \link{get_web_host_as_markdown} to get this
table as Markdown text.
}
