% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_web_host_as_markdown.R
\name{get_web_host_as_markdown}
\alias{get_web_host_as_markdown}
\title{Get all the web_host resources, as a Markdown table}
\usage{
get_web_host_as_markdown(t = get_web_host())
}
\arguments{
\item{t}{a table, as returned by \link{get_web_host}}
}
\value{
a Markdown table of all web_host resources, as text
}
\description{
This table is displayed at
\url{https://nbisweden.github.io/SCoRe_user_doc/web_host/}.
}
\seealso{
use \link{get_web_host} to get this
table as a table.
}
