% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_web_host_to_center.R
\name{get_web_host_to_center}
\alias{get_web_host_to_center}
\title{Get a table of all the web_hosts and the center(s) these
are connected to.}
\usage{
get_web_host_to_center()
}
\value{
a table with all web_hosts and the center(s) these
are connected to, in tidy data format,
sorted alphabetically by (1) web_host name,
(2) center name
}
\description{
Get a table of all the web_hosts and the center(s) these
are connected to.
}
\examples{
get_web_host_to_center()
}
