% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_correctly_formatted_date.R
\name{is_correctly_formatted_date}
\alias{is_correctly_formatted_date}
\title{Is the date correctly formatted?}
\usage{
is_correctly_formatted_date(date_str)
}
\arguments{
\item{date_str}{a string for a date}
}
\value{
TRUE if the string follows a \code{YYYY-MM-DD} format
}
\description{
Is the date correctly formatted?
}
\note{
this does not test the value of the string
}
