% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_correctly_formatted_table.R
\name{is_correctly_formatted_table}
\alias{is_correctly_formatted_table}
\title{Determine if the table is correctly formatted}
\usage{
is_correctly_formatted_table(t)
}
\description{
Determine if the table is correctly formatted
}
