% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_correctly_formatted_web_host_table.R
\name{is_correctly_formatted_web_host_table}
\alias{is_correctly_formatted_web_host_table}
\title{Determine if the table is correctly formatted}
\usage{
is_correctly_formatted_web_host_table(t)
}
\arguments{
\item{t}{a table}
}
\value{
TRUE if the table is correctly formatted
}
\description{
Determine if the table is correctly formatted
}
\examples{
is_correctly_formatted_web_host_table(get_web_host()) # True
is_correctly_formatted_web_host_table("Nonsense") # False
}
