% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mermaid_line.R
\name{is_mermaid_line}
\alias{is_mermaid_line}
\title{Is each line in a text a mermaid line?}
\usage{
is_mermaid_line(text)
}
\arguments{
\item{text}{lines of text}
}
\value{
a vector of booleans, one per line
}
\description{
Is each line in a text a mermaid line?
}
