% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_provider.R
\name{is_provider}
\alias{is_provider}
\title{Determines if the input
is the name of a provider of computational resources}
\usage{
is_provider(s)
}
\arguments{
\item{s}{the string to be tested}
}
\value{
TRUE if the name is the name of a provider
}
\description{
Determines if the input
is the name of a provider of computational resources
}
\examples{
is_provider("UPPMAX")
}
