% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_link.R
\name{is_valid_link_impl_2}
\alias{is_valid_link_impl_2}
\title{Checks if a URL is valid}
\usage{
is_valid_link_impl_2(x, non_2xx_return_value = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{x}{a single URL}

\item{non_2xx_return_value}{what to do if the site exists but the
HTTP status code is not in the \verb{2xx} range. Default is to return \code{FALSE}.}

\item{quiet}{if not \code{FALSE}, then every time the \code{non_2xx_return_value} condition
arises a warning message will be displayed. Default is \code{FALSE}.}

\item{...}{other params (\code{timeout()} would be a good one) passed directly
to \code{httr::HEAD()} and/or \code{httr::GET()}}
}
\description{
From \url{https://stackoverflow.com/a/52915256/3364162}.
}
