% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_url.R
\name{is_valid_url}
\alias{is_valid_url}
\title{Determine if the URL is correctly formatted.}
\usage{
is_valid_url(url)
}
\arguments{
\item{url}{a URL}
}
\value{
TRUE if the URL has a valid format
}
\description{
It does not check if the URL exists
}
