% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_file_as_computer_generated.R
\name{mark_file_as_computer_generated}
\alias{mark_file_as_computer_generated}
\title{Mark a filename as computer generated,
by adding comments to indicate this}
\usage{
mark_file_as_computer_generated(filename)
}
\arguments{
\item{filename}{name of the file to be changed}
}
\description{
Mark a filename as computer generated,
by adding comments to indicate this
}
