% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_urls.R
\name{parse_urls}
\alias{parse_urls}
\title{Combine a base URL with multiple relative URLs}
\usage{
parse_urls(base_url, relative_urls)
}
\arguments{
\item{base_url}{base URL}

\item{relative_urls}{a relative URL}
}
\value{
a URL
}
\description{
Combine a base URL with multiple relative URLs
}
