% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_compute_as_csv.R
\name{save_compute_as_csv}
\alias{save_compute_as_csv}
\title{Save all the courses, as a CSV text}
\usage{
save_compute_as_csv(csv_filename = "compute.csv")
}
\arguments{
\item{csv_filename}{path to the file to be created}
}
\value{
Nothing
}
\description{
Save all the courses, as a CSV text
}
