% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_courses_as_csv.R
\name{save_courses_as_csv}
\alias{save_courses_as_csv}
\title{Save all the courses, as a CSV text}
\usage{
save_courses_as_csv(csv_filename = "courses.csv")
}
\arguments{
\item{csv_filename}{path to the file to be created}
}
\value{
a Markdown table  of all courses, as text
}
\description{
Save all the courses, as a CSV text
}
