% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surround_markdown_with_lintignore.R
\name{surround_markdown_with_lintignore}
\alias{surround_markdown_with_lintignore}
\title{Surround Markdown text with Markdownlint ignore numbers}
\usage{
surround_markdown_with_lintignore(markdown_text)
}
\arguments{
\item{markdown_text}{Markdown text}
}
\value{
Markdown surrounded by Markdownlint ignores
}
\description{
Surround Markdown text with Markdownlint ignore numbers
}
