% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_english_range.R
\name{extract_english_range}
\alias{extract_english_range}
\title{Extract an English range from a piece of text}
\usage{
extract_english_range(text)
}
\arguments{
\item{text}{one line of text}
}
\value{
the English range
}
\description{
The text can be messy :-)
}
\examples{
extract_english_range("nonsense 27-28 November 2025 lalala")
}
