% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_slubi_course_names.R
\name{extract_slubi_course_names}
\alias{extract_slubi_course_names}
\title{Extract the course names from an SLUBI courses text}
\usage{
extract_slubi_course_names(course_names)
}
\arguments{
\item{slubi_courses_text}{subset of raw HTML from the SLUBI website.}
}
\description{
Extract the course names from an SLUBI courses text
}
