% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_slubi_from_date.R
\name{extract_slubi_from_date}
\alias{extract_slubi_from_date}
\title{Extract the from date from an SLUBI courses text}
\usage{
extract_slubi_from_date(english_from_date)
}
\arguments{
\item{english_from_date}{subset of raw HTML from the SLUBI website.}
}
\description{
Extract the from date from an SLUBI courses text
}
