% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_slubi_from_dates.R
\name{extract_slubi_from_dates}
\alias{extract_slubi_from_dates}
\title{Extract the from dates from an SLUBI courses text}
\usage{
extract_slubi_from_dates(english_from_dates)
}
\arguments{
\item{english_from_dates}{subset of raw HTML from the SLUBI website.}
}
\description{
Extract the from dates from an SLUBI courses text
}
