% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_slubi_to_date.R
\name{extract_slubi_to_date}
\alias{extract_slubi_to_date}
\title{Extract the to date from an SLUBI courses text}
\usage{
extract_slubi_to_date(english_date_range)
}
\arguments{
\item{english_date_range}{a data range in English}
}
\description{
Extract the to date from an SLUBI courses text
}
