% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adh_courses.R
\name{get_adh_courses}
\alias{get_adh_courses}
\title{Get the AIDA Data Hub courses}
\usage{
get_adh_courses()
}
\value{
a table with all AIDA Data Hub courses.
}
\description{
Get the AIDA Data Hub courses
}
