% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_provider_logos_as_markdown.R
\name{get_all_provider_logos_as_markdown}
\alias{get_all_provider_logos_as_markdown}
\title{Get the logos of all the providers as Markdown}
\usage{
get_all_provider_logos_as_markdown()
}
\value{
the logos of all the providers as Markdown
}
\description{
Get the logos of all the providers as Markdown
}
\examples{
get_all_provider_logos_as_markdown()
}
\seealso{
use \link{get_provider_logo_as_markdown} to get the
Markdown of a single clickable logo
}
