% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_courses.R
\name{get_courses}
\alias{get_courses}
\title{Get a table with all of the future courses}
\usage{
get_courses()
}
\value{
a table with all future courses, sorted alphabetically by date
}
\description{
Get a table with all of the future courses
}
