% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hpc2n_test_courses_text.R
\name{get_hpc2n_test_courses_text}
\alias{get_hpc2n_test_courses_text}
\title{Get a testing text for HPC2N courses}
\usage{
get_hpc2n_test_courses_text()
}
\description{
Get a testing text for HPC2N courses
}
