% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infraviz_courses.R
\name{get_infraviz_courses}
\alias{get_infraviz_courses}
\title{Get the InfraViz courses}
\usage{
get_infraviz_courses()
}
\value{
a table with all InfraViz courses.
}
\description{
Get the InfraViz courses
}
