% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naiss_courses.R
\name{get_naiss_courses}
\alias{get_naiss_courses}
\title{Get the NAISS courses}
\usage{
get_naiss_courses()
}
\value{
a table with all NAISS courses.

This table is a \link[tibble]{tibble} with these column names:
\itemize{
\item \code{source}: the URL where this course was found
\item \code{name}: the name of the course
\item \code{url}: the URL of the course
}
}
\description{
Get the NAISS courses
}
