% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naiss_courses.R
\name{get_naiss_courses}
\alias{get_naiss_courses}
\title{Get the NAISS courses}
\usage{
get_naiss_courses(html_text = scoreto::get_naiss_html())
}
\arguments{
\item{html_text}{HTML text to parse, as can be obtained by
\link{get_naiss_html} or \link{get_test_naiss_html}}
}
\value{
a table with all NAISS courses, where
the table will pass the test of
\link{is_correctly_formatted_courses_table}
}
\description{
Get the NAISS courses
}
