% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nsc_courses.R
\name{get_nsc_courses}
\alias{get_nsc_courses}
\title{Get the NSC courses}
\usage{
get_nsc_courses(html_text = scoreto::get_nsc_html())
}
\value{
a table with all NSC courses.
}
\description{
Get the NSC courses
}
