% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sll_courses.R
\name{get_sll_courses}
\alias{get_sll_courses}
\title{Get the SciLifeLab courses}
\usage{
get_sll_courses(html_text = scoreto::get_sll_html())
}
\value{
a table with all SciLifeLab courses.
}
\description{
Get the SciLifeLab courses
}
