% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slubi_courses.R
\name{get_slubi_courses}
\alias{get_slubi_courses}
\title{Get the SLUBI courses}
\usage{
get_slubi_courses(html_text = scoreto::get_slubi_html())
}
\value{
a table with all SLUBI courses.
}
\description{
Get the SLUBI courses
}
