% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storage_to_center.R
\name{get_storage_to_center}
\alias{get_storage_to_center}
\title{Get a table of all the HPC storage systems and the center(s) these
are connected to.}
\usage{
get_storage_to_center()
}
\value{
a table with all HPC storage systems and the center(s) these
are connected to, in tidy data format,
sorted alphabetically by (1) HPC storage system name,
(2) HPC center name
}
\description{
Get a table of all the HPC storage systems and the center(s) these
are connected to.
}
\examples{
get_storage_to_center()
}
